refactor(jsx): move all shared functions to chartUtils.js
