+++ fullstack/src/components/graphs/LineChartComponent.jsx
@@ -8,43 +8,11 @@ import {
   Tooltip,
   ResponsiveContainer,
 } from 'recharts';
-import { calculateYAxisConfig } from '../../app/utils/chartUtils';
-import { parseISO, format } from 'date-fns';
-
-// Custom tick component for displaying date and day
-const CustomXAxisTick = ({ x, y, payload }) => {
-  try {
-    const date = parseISO(payload.value);
-    return (
-      <g transform={`translate(${x},${y})`}>
-        <text
-          x={0}
-          y={0}
-          dy={10}
-          textAnchor="end"
-          fill="#113f67"
-          fontSize={10}
-          transform="rotate(-45)"
-        >
-          <tspan x={0} dy="1em">
-            {format(date, 'dd/MM')}
-          </tspan>
-          <tspan x={0} dy="1em">
-            {format(date, 'EEE')}
-          </tspan>
-        </text>
-      </g>
-    );
-  } catch (error) {
-    console.error(
-      'Error formatting date:',
-      error,
-      'Original value:',
-      payload.value,
-    );
-    return null;
-  }
-};
+import {
+  calculateYAxisConfig,
+  filterAndSortData,
+  CustomXAxisTick,
+} from '../../app/utils/chartUtils';
 
 const LineChartComponent = ({ data, datakey, viewType }) => {
   const [isScrollEnabled, setIsScrollEnabled] = useState(
@@ -62,40 +30,16 @@ const LineChartComponent = ({ data, datakey, viewType }) => {
   }, []);
 
   const { domain, ticks } = calculateYAxisConfig(data, datakey);
-
   const xAxisDataKey = viewType === 'hourly' ? 'hour' : 'day';
 
-  // Ensure data is valid and sorted
-  const dateKey = viewType === 'hourly' ? 'hour' : 'day';
-  const validData = data.filter(
-    (item) => item[dateKey] !== undefined && item[dateKey] !== null,
-  );
-
-  const sortedData = validData.slice().sort((a, b) => {
-    const dateA = parseISO(a[dateKey]);
-    const dateB = parseISO(b[dateKey]);
-    return isNaN(dateA) || isNaN(dateB) ? 0 : dateA - dateB;
-  });
-
-  // Slice data based on the viewType without changing its order
-  const filteredData = sortedData.slice(viewType === 'hourly' ? -24 : -30);
+  // Use the utility function for filtering and sorting data
+  const filteredData = filterAndSortData(data, xAxisDataKey, viewType);
 
   // Adjust container width for scroll functionality
   const containerWidth =
     viewType !== '7days' && isScrollEnabled
       ? `${Math.max(filteredData.length * 50, window.innerWidth)}px`
       : '100%';
-
-  const formatXAxis = (tick) => {
-    if (!tick) return 'No Data';
-    try {
-      const date = parseISO(tick);
-      return `${format(date, 'dd/MM')} (${format(date, 'EEE')})`;
-    } catch (error) {
-      console.error('Error formatting date:', error, 'Original tick:', tick);
-      return 'Invalid Date';
-    }
-  };
 
   return (
     <div
@@ -119,7 +63,6 @@ const LineChartComponent = ({ data, datakey, viewType }) => {
               stroke={graphColor}
               tick={<CustomXAxisTick />}
               tickLine={{ transform: 'translateY(5px)' }}
-              tickFormatter={formatXAxis}
               textAnchor="end"
               angle={-45}
               dy={10}
@@ -145,7 +88,7 @@ const LineChartComponent = ({ data, datakey, viewType }) => {
             />
             <Line
               type="monotone"
-              dataKey="avg_value"
+              dataKey={datakey}
               stroke={graphColor}
               strokeWidth={2}
               dot={{ fill: graphColor, r: 3 }}
